Added functionality Telegram Bot

+++ src/modules/cron/cron.service.ts
@@ -3,44 +3,62 @@ import { Injectable } from '@nestjs/common';
 import { MailService } from '../libs/mail/mail.service';
 import { Cron } from '@nestjs/schedule';
 import { StorageService } from '../libs/storage/storage.service';
+import { TelegramService } from '../libs/telegram/telegram.service';
 
 @Injectable()
 export class CronService {
-    public  constructor(
+    public constructor(
         private readonly prismaService: PrismaService,
         private readonly mailService: MailService,
-        private readonly storageService: StorageService
-    ) {}
+        private readonly storageService: StorageService,
+        private readonly telegramService: TelegramService
+    ) { }
 
     //@Cron('*/10 * * * * *') 
-    @Cron('0 0 * * *') 
+    @Cron('0 0 * * *')
     public async deleteDeactivatedAccounts() {
 
-       const sevenDaysAgo = new Date()
-       sevenDaysAgo.setDate(sevenDaysAgo.getDay() - 7)
-       //sevenDaysAgo.setDate(sevenDaysAgo.getDay() - 5)
+        const sevenDaysAgo = new Date()
+        sevenDaysAgo.setDate(sevenDaysAgo.getDay() - 7)
+        //sevenDaysAgo.setDate(sevenDaysAgo.getDay() - 5)
 
-       const deactivatedAccounts = await this.prismaService.user.findMany({
-        where: {
-            isDeactivated: true,
-            deactivatedAt: {
-                lte: sevenDaysAgo
+        const deactivatedAccounts = await this.prismaService.user.findMany({
+            where: {
+                isDeactivated: true,
+                deactivatedAt: {
+                    lte: sevenDaysAgo
+                }
+            },
+            include: {
+                notificationSettings: true,
+                stream: true
             }
-        }
-       })
-       for(const user of deactivatedAccounts) {
-        await this.mailService.sendAccountDeletion(user.email)
+        })
+        for (const user of deactivatedAccounts) {
+            await this.mailService.sendAccountDeletion(user.email)
 
-        this.storageService.remove(user.avatar)
-       }
+            if (user.notificationSettings.telegramNotifications && user.telegramId) {
 
-       await this.prismaService.user.deleteMany({
-        where: {
-            isDeactivated: true,
-            deactivatedAt: {
-                lte: sevenDaysAgo
+                await this.telegramService.sendAccountDeletion(user.telegramId)
             }
+
+            if (user.avatar) {
+                this.storageService.remove(user.avatar)
+            }
+
+            if (user.stream.thumbnailUrl) {
+                this.storageService.remove(user.stream.thumbnailUrl)
+            }
+
         }
-       })
+
+        await this.prismaService.user.deleteMany({
+            where: {
+                isDeactivated: true,
+                deactivatedAt: {
+                    lte: sevenDaysAgo
+                }
+            }
+        })
     }
 }
